Fix crash with Int overflow of rx field.

+++ src/main/kotlin/nl/smeh/vpnman/RouterOsClient.kt
@@ -49,8 +49,8 @@ data class InterfaceWireguardPeers(
     @JsonProperty("current-endpoint-address") val currentEndpointAddress: String? = null,
     @JsonProperty("current-endpoint-port") val currentEndpointPort: Int? = null,
     @JsonProperty("last-handshake") val lastHandshake: String? = null,
-    @JsonProperty("rx") val rx: Int? = null,
-    @JsonProperty("tx") val tx: Int? = null,
+    @JsonProperty("rx") val rx: Long? = null,
+    @JsonProperty("tx") val tx: Long? = null,
 )
 
 @ResourcePath("/routing/table")
